--- MAJOR UPDATE ---
- Updated ImGui for Docking
- Added RenderTarget for camera view port
- Debug Log
- Shader Update

+++ GraphicsEngine3D/UBaseObject.h
@@ -30,7 +30,7 @@ protected:
 	bool m_Valid = true;
 	virtual void OnDestroyed() {};
 
-	size_t ID;
+	int32 ID;
 
 public:
 	World* GetWorld() const;
